refactor: 입력 처리 함수를 별도 파일로 분리
